removing warnings under ubuntu

+++ src/addons/Engineering/Mechanics/VRGearSegmentation.cpp
@@ -440,8 +440,8 @@ void VRGearSegmentation::computeGearParams(int fN) {
         VertexPlane& plane1 = planes[match.first];
         VertexPlane& plane2 = planes[match.second];
 
-        if (plane1.sineFits.size() <= fN) continue;
-        if (plane2.sineFits.size() <= fN) continue;
+        if (plane1.sineFits.size() <= size_t(fN)) continue;
+        if (plane2.sineFits.size() <= size_t(fN)) continue;
         if (plane1.sineFits[fN].sineFitParams.size() == 0) continue;
         if (plane2.sineFits[fN].sineFitParams.size() == 0) continue;
 
@@ -488,12 +488,12 @@ void VRGearSegmentation::analyse(VRObjectPtr o) {
 
 Vec3d VRGearSegmentation::getAxis() { return axis; }
 PosePtr VRGearSegmentation::getPolarCoords() { return Pose::create(Vec3d(0,0,0), axis, r1); }
-int VRGearSegmentation::getNGears() { return gears.size(); }
-int VRGearSegmentation::getNPlanes() { return planes.size(); }
+size_t VRGearSegmentation::getNGears() { return gears.size(); }
+size_t VRGearSegmentation::getNPlanes() { return planes.size(); }
 
-vector<double> VRGearSegmentation::getGearParams(int i) { return i<gears.size() ? gears[i] : vector<double>(); }
+vector<double> VRGearSegmentation::getGearParams(size_t i) { return i<gears.size() ? gears[i] : vector<double>(); }
 
-VRGeometryPtr VRGearSegmentation::createGear(int i) {
+VRGeometryPtr VRGearSegmentation::createGear(size_t i) {
     auto p = getGearParams(i);
     if (p.size() < 8) return 0;
     auto g = VRGeometry::create("gear");
@@ -503,20 +503,20 @@ VRGeometryPtr VRGearSegmentation::createGear(int i) {
     return g;
 }
 
-double VRGearSegmentation::getPlanePosition(int i) { return planes[i].position; }
-vector<Vec2d> VRGearSegmentation::getPlaneContour(int i) { return planes[i].contour; }
+double VRGearSegmentation::getPlanePosition(size_t i) { return planes[i].position; }
+vector<Vec2d> VRGearSegmentation::getPlaneContour(size_t i) { return planes[i].contour; }
 
-vector<double> VRGearSegmentation::getPlaneSineGuess(int i, int sf) {
+vector<double> VRGearSegmentation::getPlaneSineGuess(size_t i, size_t sf) {
     if (planes[i].sineFits.size() <= sf) return vector<double>();
     return planes[i].sineFits[sf].guess;
 }
 
-vector<double> VRGearSegmentation::getPlaneSineApprox(int i, int sf) {
+vector<double> VRGearSegmentation::getPlaneSineApprox(size_t i, size_t sf) {
     if (planes[i].sineFits.size() <= sf) return vector<double>();
     return planes[i].sineFits[sf].sineFitParams;
 }
 
-vector<Vec2d> VRGearSegmentation::getPlaneVertices(int i) {
+vector<Vec2d> VRGearSegmentation::getPlaneVertices(size_t i) {
     vector<Vec2d> vp;
     for (auto& v : planes[i].vertices) vp.push_back(v.polarCoords);
     return vp;
